move gengapic to internal
The interface isn't stable and we don't want to advertise it yet.
We can always move it back when we want to.
