+++ internal/gengapic/example.go
