+++ internal/gengapic/stream.go
