+++ internal/gengapic/testdata/doc_file.want
