+++ internal/gengapic/testdata/empty_client_init.want
