+++ internal/gengapic/testdata/empty_example.want
