+++ internal/gengapic/testdata/empty_opt.want
