+++ internal/gengapic/testdata/foo_client_init.want
