+++ internal/gengapic/testdata/foo_example.want
