+++ internal/gengapic/testdata/foo_opt.want
