+++ internal/gengapic/testdata/method_BidiThings.want
