+++ internal/gengapic/testdata/method_GetBigThing.want
