+++ internal/gengapic/testdata/method_GetEmptyThing.want
