+++ internal/gengapic/testdata/method_GetManyThings.want
