+++ internal/gengapic/testdata/method_GetOneThing.want
